LayoutTests/http/tests/xmlhttprequest/access-control-basic-denied-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-denied.cgi. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [NETWORK_ERR: XMLHttpRequest Exception 101].
